Último commit con código grupal

+++ src/main/java/com/FinalEgg/ServiChacras/entidades/Proveedor.java
@@ -37,5 +37,5 @@ public class Proveedor {
     private Integer cantPedido;
 
     @Column(name = "promedio_puntuacion")
-    private Float promPuntuacion;
+    private Double promPuntuacion;
 }
